+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomAvatarIndex, getRandomLocationX, getRandomLocationY, getRandomNumbers, getRandomElements} from './util.js';
+
 const HEADERS = [
   'Шикарный дом на берегу залива',
   'Уютная квартира для молодой пары',
@@ -55,4 +57,43 @@ const PHOTOS = [
 
 const SIMILAR_AD_AMOUNT = 10;
 
+
+/**
+ * Функция генерирования объекта
+ *  @return {object} — объект объявления
+ */
+function createAd() {  // const createAd = () => {
+
+  const coordinateX = getRandomLocationX(35.65000, 35.70000, 5);
+  const coordinateY = getRandomLocationY(139.70000, 139.80000, 5);
+  const addressFull = coordinateX + ', ' + coordinateY;
+
+  return {
+    author: {
+      avatar: 'img/avatars/users0' + getRandomAvatarIndex(1, 8) + '.png',
+    },
+    offer: {
+      title: getRandomElements(HEADERS),
+      address: addressFull,
+      price: getRandomNumbers(1, 50000),
+      type: getRandomElements(TYPES),
+      rooms: getRandomNumbers(1, 20),
+      guests: getRandomNumbers(1, 20),
+      checkin: getRandomElements(CHECKTIME),
+      checkout: getRandomElements(CHECKTIME),
+      features: getRandomElements(FEATURES),
+      description: getRandomElements(DESCRIPTIONS),
+      photos: getRandomElements(PHOTOS),
+    },
+    location: {
+      x: coordinateX,
+      y: coordinateY,
+    },
+  }
+}
+
+const similarAds = new Array(SIMILAR_AD_AMOUNT).fill(1).map(() => createAd());
+
+export {similarAds}
+export {createAd}
 export {HEADERS, TYPES, CHECKTIME, FEATURES, DESCRIPTIONS, PHOTOS, SIMILAR_AD_AMOUNT}
