+++ js/util.js
@@ -1,5 +1,3 @@
-import {HEADERS, TYPES, CHECKTIME, FEATURES, DESCRIPTIONS, PHOTOS} from './base.js';
-
 /**
  * Функция получения случайного ЦЕЛОГО числа для задания номера аватара
  * @param {number} min — минимальное значение
@@ -57,41 +55,4 @@ const getRandomElements = (sets) => {
   return sets[Math.floor(Math.random() * sets.length)];
 }
 
-
-
-/**
- * Функция генерирования объекта
- *  @return {object} — объект объявления
- */
-function createAd() {  // const createAd = () => {
-
-  const coordinateX = getRandomLocationX(35.65000, 35.70000, 5);
-  const coordinateY = getRandomLocationY(139.70000, 139.80000, 5);
-  const addressFull = coordinateX + ', ' + coordinateY;
-
-  return {
-    author: {
-      avatar: 'img/avatars/users0' + getRandomAvatarIndex(1, 8) + '.png',
-    },
-    offer: {
-      title: getRandomElements(HEADERS),
-      address: addressFull,
-      price: getRandomNumbers(1, 50000),
-      type: getRandomElements(TYPES),
-      rooms: getRandomNumbers(1, 20),
-      guests: getRandomNumbers(1, 20),
-      checkin: getRandomElements(CHECKTIME),
-      checkout: getRandomElements(CHECKTIME),
-      features: getRandomElements(FEATURES),
-      description: getRandomElements(DESCRIPTIONS),
-      photos: getRandomElements(PHOTOS),
-    },
-    location: {
-      x: coordinateX,
-      y: coordinateY,
-    },
-  }
-}
-
-export {createAd}
-
+export {getRandomAvatarIndex, getRandomLocationX, getRandomLocationY, getRandomNumbers, getRandomElements}
